Lazy-load panorama viewer
